+++ packages/generator/src/parsers/extractQualifier.ts
@@ -1,0 +1,17 @@
+import { FULLY_QUALIFIED, MINIMALLY_QUALIFIED, Qualifier, UNQUALIFIED } from 'emojibase';
+
+export function extractQualifier(type: string): Qualifier | null {
+	if (type === 'fully-qualified') {
+		return FULLY_QUALIFIED;
+	}
+
+	if (type === 'minimally-qualified') {
+		return MINIMALLY_QUALIFIED;
+	}
+
+	if (type === 'unqualified') {
+		return UNQUALIFIED;
+	}
+
+	return null;
+}